🚨 Change int to double in power/health data members

+++ Alien-Invasion/Units/AlienDrone.cpp
@@ -1,7 +1,7 @@
 #include"AlienDrone.h"
 #include"../Game class/Game.h"
 
-AlienDrone::AlienDrone(int H, int P, int AC, int tj, Game* Gameptr):Unit(H,P,AC,tj, Gameptr)
+AlienDrone::AlienDrone(double H, double P, int AC, int tj, Game* Gameptr):Unit(H,P,AC,tj, Gameptr)
 {
 	Type = alienDrone;
 }
@@ -42,7 +42,7 @@ void AlienDrone::Attack()
 			unit2->setTa(pGame->GetCurrentTime()); //Set Ta (first attacked time)
 
 
-			int Damage = (this->getHealth() * this->getPower() / 100) /
+			double Damage = (this->getHealth() * this->getPower() / 100) /
 				sqrt(unit2->getHealth());	//Damage Formula
