+++ Alien-Invasion/Units/EarthSoldier.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 #include "Earthsoldier.h"
 #include"../Game Class/Game.h"
-EarthSoldier::EarthSoldier(int H, int P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
+EarthSoldier::EarthSoldier(double H, double P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
 {
 	Type = earthSoldier;
 }
@@ -25,7 +25,7 @@ void EarthSoldier::Attack()
 			unit2->setTa(pGame->GetCurrentTime()); //Set Ta (first attacked time)
 
 
-			int Damage = (this->getHealth() * this->getPower() / 100) /
+			double Damage = (this->getHealth() * this->getPower() / 100) /
 				sqrt(unit2->getHealth());	//Damage Formula
