+++ Alien-Invasion/Units/SaverUnit.cpp
@@ -1,7 +1,7 @@
 #include "SaverUnit.h"
 #include"../Game Class/Game.h"
 
-SaverUnit::SaverUnit(int H, int P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
+SaverUnit::SaverUnit(double H, double P, int AC, int tj, Game* Gameptr) :Unit(H, P, AC, tj, Gameptr)
 {
     Type = saverUnit;
 }
@@ -23,7 +23,7 @@ void SaverUnit::Attack()
             unit2->setTa(pGame->GetCurrentTime()); //Set Ta (first attacked time)
 
 
-            int Damage = (this->getHealth() * this->getPower() / 100) /
+            double Damage = (this->getHealth() * this->getPower() / 100) /
                 sqrt(unit2->getHealth());	//Damage Formula
